Refaktoriral kodo iz utils.py nazaj v podatki.py.
